添加视频

+++ iiiviviniko/admin/ajaxoperation.php
@@ -6,22 +6,41 @@ if($method == "delpage"){
 	$id = $_GET['id'];
 	$pagedb = new Page($dbutil);
 	$page = $pagedb->getPageById($id);
-	$img_s = $_SERVER['DOCUMENT_ROOT'].$page->imgsmall;
-	$img_b = $_SERVER['DOCUMENT_ROOT'].$page->imgbig;
+	$img_s = $_SERVER['DOCUMENT_ROOT'].$root_path.$page->imgsmall;
 	if(file_exists($img_s)){
 		unlink($img_s);
 	}
-	if(file_exists($img_b)){
-		unlink($img_b);
+	if($page->isvideo == 1){
+		$img_b = $_SERVER['DOCUMENT_ROOT'].$root_path.$page->imgbig;
+		if(file_exists($img_b)){
+			unlink($img_b);
+		}	
 	}
 	$pagedb->deletePageById($id);
 	echo 'success';
 }else if($method == "delcate"){
 	$id = $_GET['id'];
-	
+	$pagedb = new Page($dbutil);
+	$pages = $pagedb->getPagesAllByPid($id);
+	if(count($pages) >0){
+		foreach ($pages as $page){
+			$img_s = $_SERVER['DOCUMENT_ROOT'].$root_path.$page->imgsmall;
+			echo $img_s;
+			if(file_exists($img_s)){
+				unlink($img_s);
+			}
+			if($page->isvideo != 1){
+				$img_b = $_SERVER['DOCUMENT_ROOT'].$root_path.$page->imgbig;
+				if(file_exists($img_b)){
+					unlink($img_b);
+				}	
+			}
+		}
+		$pagedb->deletePageByPId($id);
+	}
 	$catedb = new Category($dbutil);
 	$page = $catedb->getMenuInfoById($id);
-	$img_s = $_SERVER['DOCUMENT_ROOT'].$page['img'];
+	$img_s = $_SERVER['DOCUMENT_ROOT'].$root_path.$page['img'];
 	if(file_exists($img_s)){
 		unlink($img_s);
 	}
